don't even remember
what i updated heh

+++ emulators/FnkBstrServices/FnkBstrA/FnkBstrA.cs
@@ -7,20 +7,35 @@ using SocketService;
 using System.Net.Sockets;
 using Microsoft.Win32;
 using System.Security.Cryptography;
+using System.Diagnostics;
 
 
 namespace FnkBstrA
 {
-    class FnkBstrA : UDPSocketService
+    class FnkBstrA : UDPServerService
     {
-        protected int _version = 1033;
-        protected const string pnkBstrASubKey = "SOFTWARE\\Even Balance\\";
+        public class ProxyData
+        {
+            public string pnkBstrBSubKey = "SOFTWARE\\Even Balance\\";
+            public string pnkBstrBserviceKey = "PnkBstrB";
+            public string pnkBstrBWow64SubKey = "SOFTWARE\\Wow6432Node\\Even Balance\\";
+            public int source_port = 45301;
+            public int destination_port = 45308;
+            public string proxy_app = "UDPDP.exe";
+            public string log_file = "pnkbstrb_packets.log";
+            public string host = "127.0.0.1";
+
+        }
+        protected int _version = 1036;
+        protected int _port = 44301;
         protected string _pnkbstrPath = "C:\\Tools\\";
+        protected const string pnkBstrASubKey = "SOFTWARE\\Even Balance\\";
+        protected const string pnkBstrAWow64SubKey = "SOFTWARE\\Wow6432Node\\Even Balance\\";
         protected const string serviceKey = "PnkBstrA";
         protected const string returnOk = "l0";
         protected const string returnFail = "l1";
         protected const string returnTooManyLoads = "l2";
-
+        protected ProxyData proxy = null;
         public FnkBstrA( int port )
             : base ( 44301 )
         {
@@ -41,10 +56,107 @@ namespace FnkBstrA
             _version = version;
             _pnkbstrPath = path;
         }
+        public void ConfigureProxy()
+        {
+            // 45301 (default pnkbstrb port
+            proxy = new ProxyData();
+        }
+        // reads in the port value that PnkBstrB set, then overwrites it
+        // with our proxy port.
+        public bool StartProxy()
+        {
+            try
+            {
+                LaunchProxyApp();
+                System.Threading.Thread.Sleep(500);
+                return true;
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine("Error starting proxy service: {0}", e.Message);
+                proxy = null;
+            }
+            return false;
+        }
+
+        public void OverwritePnkBstrBPort()
+        {
+            try
+            {
+                System.Threading.Thread.Sleep(200);
+                var hEvenBalanceKey = Registry.LocalMachine.CreateSubKey(proxy.pnkBstrBSubKey);
+                var hServiceKey = hEvenBalanceKey.CreateSubKey(proxy.pnkBstrBserviceKey);
+                var hWowEvenBalanceKey = Registry.LocalMachine.CreateSubKey(proxy.pnkBstrBWow64SubKey);
+                var hWowServiceKey = hWowEvenBalanceKey.CreateSubKey(proxy.pnkBstrBserviceKey);
+                proxy.destination_port = (int)hWowServiceKey.GetValue("Port");
+
+                Console.WriteLine("PnkBstrB wrote port: {0} to the registry, overwriting with {1}...", proxy.destination_port, proxy.source_port);
+                hWowServiceKey.SetValue("Port", proxy.source_port);
+                hServiceKey.SetValue("Port", proxy.source_port);
+                Console.WriteLine("Wrote Port: {0} to registry.", proxy.source_port);
+            }
+            catch (Exception e)
+            {
+                Console.WriteLine("Error writing port to registry for proxy hijack: {0}", e.Message);
+                proxy = null;
+            }
+        }
+
+        private void LaunchProxyApp()
+        {
+            string args = string.Format("-s {0}:{1} -d {0}:{3} -o {4} -f raw", proxy.host, proxy.source_port, proxy.host, proxy.destination_port, proxy.log_file);
+            ProcessStartInfo startInfo = new ProcessStartInfo();
+            startInfo.CreateNoWindow = true;
+            startInfo.UseShellExecute = true;
+            startInfo.FileName = proxy.proxy_app;
+            startInfo.WorkingDirectory = "C:\\tools\\";
+            Console.WriteLine("Starting UDPDP.exe {0}", args);
+            startInfo.Arguments = args;
+            Process p = Process.Start(startInfo);
+            
+        }
+        public void InitPnkBstrB(string installCmd)
+        {
+            string pnkbstrInstallCmd = installCmd + _pnkbstrPath + "PnkBstrB.exe";
+            UninstallPnkBstr("PnkBstrB");
+            InstallPnkBstr(pnkbstrInstallCmd, "PnkBstrB");
+            return;
+        }
 
         public string[] ParseMemoryCommand( string command )
         {
-            return command.ToString ( ).Split ( ' ' );
+            return command.ToString( ).Split( ' ' );
+        }
+
+        static void InstallPnkBstr( string pnkbstrInstallCmd, string serviceName )
+        {
+            if (ServiceTools.ServiceInstaller.ServiceIsInstalled(serviceName))
+            {
+                Console.WriteLine("Service is already installed!");
+                return;
+            }
+            try
+            {
+                ServiceTools.ServiceInstaller.Install(serviceName, serviceName, pnkbstrInstallCmd);
+                Console.WriteLine("PnkBstrB Service Installed as: ");
+                Console.WriteLine(pnkbstrInstallCmd);
+
+            }
+            catch (System.ApplicationException e)
+            {
+                Console.WriteLine("Error installing service (do you have proper privileges?): {0}", e.Message);
+            }
+        }
+
+        static void UninstallPnkBstr(string serviceName)
+        {
+            if (!ServiceTools.ServiceInstaller.ServiceIsInstalled(serviceName))
+            {
+                Console.WriteLine("Service is already uninstalled!");
+                return;
+            }
+            ServiceTools.ServiceInstaller.Uninstall(serviceName);
+            Console.WriteLine("Service Uninstalled!");
         }
 
         public void SetRegistryKeys( )
@@ -53,8 +165,13 @@ namespace FnkBstrA
             {
                 var hEvenBalanceKey = Registry.LocalMachine.CreateSubKey ( pnkBstrASubKey );
                 var hServiceKey = hEvenBalanceKey.CreateSubKey ( serviceKey );
+                var hWowEvenBalanceKey = Registry.LocalMachine.CreateSubKey(pnkBstrAWow64SubKey);
+                var hWowServiceKey = hWowEvenBalanceKey.CreateSubKey(serviceKey);
                 hServiceKey.SetValue ( "Port", _port );
                 hServiceKey.SetValue ( "Version", _version );
+                hWowServiceKey.SetValue("Port", _port);
+                hWowServiceKey.SetValue("Version", _version);
+                Console.WriteLine("Wrote Port: {0} and Version: {1} to registry.", _port, _version);
             }
             catch ( Exception e )
             {
@@ -73,8 +190,11 @@ namespace FnkBstrA
                 {
                     DateTime fileCreatedDate = File.GetCreationTime ( pnkbstrFile );
                     backupFile = string.Format( "{0}PnkBstrB.{1}-{2}-{3}.exe", _pnkbstrPath, fileCreatedDate.Year, fileCreatedDate.Month, fileCreatedDate.Day );
-                    File.Move ( pnkbstrFile, backupFile );
-                    Console.WriteLine( "Moved old PnkBstrB {0} to %{1}", pnkbstrFile, backupFile );
+                    if (!File.Exists(backupFile))
+                    {
+                        File.Move(pnkbstrFile, backupFile);
+                        Console.WriteLine("Moved old PnkBstrB {0} to {1}", pnkbstrFile, backupFile);
+                    }
                 }
                 catch ( Exception ex )
                 {
@@ -83,7 +203,7 @@ namespace FnkBstrA
             }
             try
             {
-                File.Copy ( new_path, pnkbstrFile );
+                File.Copy ( new_path, pnkbstrFile, true );
                 Console.WriteLine( "Copied new file {0} to {1}", new_path, pnkbstrFile );
                 return true;
             }
@@ -101,7 +221,7 @@ namespace FnkBstrA
             {
                 byte[] new_hash = MD5.Create().ComputeHash( File.ReadAllBytes( new_path ) );
                 byte[] old_hash = MD5.Create().ComputeHash( File.ReadAllBytes( pnkbstrbFile ) );
-                if ( new_hash == old_hash )
+                if ( new_hash.SequenceEqual(old_hash) )
                 {
                     return false;
                 }
@@ -124,13 +244,23 @@ namespace FnkBstrA
             {
                 case 'l':
                     Console.WriteLine ( "Client sent a service load command: {0} copying & starting PnkBstrB service.", state.sb.ToString ( ) );
-                    // Remove leading 'l' command character.
-                    string newPnkbstrFile = state.sb.ToString().Substring( 1 );
+                    // Remove leading 'l' command character and make sure no nulls exist.
+                    string newPnkbstrFile = state.sb.ToString().Substring(1).Replace("\0", string.Empty); ;
                     if ( IsNewPnkBstrB( newPnkbstrFile ) )
                     {
                         CopyNewPnkBstrB( newPnkbstrFile );
                     }
-                    ServiceTools.ServiceInstaller.StartService ( "PnkBstrB" );
+
+                    if (proxy != null)
+                    {
+                        StartProxy();
+                        ServiceTools.ServiceInstaller.StartService("PnkBstrB");
+                        OverwritePnkBstrBPort();
+                    }
+                    else
+                    {
+                        ServiceTools.ServiceInstaller.StartService("PnkBstrB");
+                    }
                     break;
                 case 'm':
                     try
@@ -155,10 +285,14 @@ namespace FnkBstrA
                     sendState.buffer = Encoding.ASCII.GetBytes ( string.Format ( "v{0}", _version ) );
                     break;
             }
-            sendState.workSocket = new Socket ( AddressFamily.InterNetwork,
+            sendState.workSocket = new Socket(AddressFamily.InterNetwork,
                                                SocketType.Dgram,
-                                               ProtocolType.Udp );
+                                               ProtocolType.Udp);
             SendData ( sendState, new AsyncCallback ( OnSent ) );
+        }
+        ~FnkBstrA()
+        {
+            UninstallPnkBstr("PnkBstrB");
         }
     }
 }
